# docs/swagger.yaml
basePath: /
definitions:
...
  goals.Range:
    properties:
      max:
        type: number
      min:
        type: number
    type: object
...
  meals.MealEntry:
    properties:
      created_at:
        type: string
      effective_name:
        type: string
      effective_nutriments_per_100g:
        $ref: '#/definitions/meals.Nutriments'
      id:
        type: string
      logged_at:
        type: string
      meal_type:
        $ref: '#/definitions/meals.MealType'
      note:
        type: string
      product_id:
        type: string
      quantity_g:
        type: number
      updated_at:
        type: string
      workout_id:
        type: string
    type: object
  meals.MealType:
    enum:
    - breakfast
    - lunch
    - dinner
    - snack
    type: string
    x-enum-varnames:
    - Breakfast
    - Lunch
    - Dinner
    - Snack
  meals.Nutriments:
    properties:
      calcium_mg:
        type: number
      carbs_g:
        type: number
      fat_g:
        type: number
      fiber_g:
        type: number
      iron_mg:
        type: number
      kcal:
        type: number
      magnesium_mg:
        type: number
      potassium_mg:
        type: number
      protein_g:
        type: number
      salt_g:
        type: number
      sugar_g:
        type: number
      vitamin_b12_mcg:
        type: number
      vitamin_c_mg:
        type: number
      vitamin_d_mcg:
        type: number
      zinc_mg:
        type: number
    type: object
...
  summary.Adherence:
    additionalProperties:
      $ref: '#/definitions/summary.AdherenceEntry'
    type: object
  summary.AdherenceEntry:
    properties:
      actual:
        type: number
      delta_pct:
        type: number
      status:
        description: under | on | over | no_data
        type: string
      target:
        $ref: '#/definitions/goals.Range'
    type: object
  summary.Daily:
    properties:
      adherence:
        $ref: '#/definitions/summary.Adherence'
      date:
        type: string
      entries:
        items:
          $ref: '#/definitions/meals.MealEntry'
        type: array
      goal_source:
        type: string
      meal_type:
        type: string
      phase_name:
        type: string
      totals:
        $ref: '#/definitions/summary.Totals'
      tz:
        type: string
    type: object
  summary.ProteinDistribution:
    properties:
      body_weight_kg:
        type: number
      body_weight_source:
        type: string
      date:
        type: string
      meal_count:
        type: integer
      meals:
        items:
          $ref: '#/definitions/summary.ProteinMeal'
        type: array
      mps_effective_meal_count:
        type: integer
      mps_threshold_g:
        type: number
      total_protein_g:
        type: number
      tz:
        type: string
    type: object
  summary.ProteinMeal:
    properties:
      gap_minutes_since_previous:
        type: integer
      logged_at:
        type: string
      logged_at_hour:
        type: integer
      meal_type:
        type: string
      mps_effective:
        type: boolean
      protein_g:
        type: number
    type: object
  summary.Range:
    properties:
      days:
        items:
          $ref: '#/definitions/summary.RangeDay'
        type: array
      from:
        type: string
      group_by:
        type: string
      to:
        type: string
      tz:
        type: string
    type: object
  summary.RangeDay:
    properties:
      adherence:
        $ref: '#/definitions/summary.Adherence'
      by_meal_type:
        additionalProperties:
          $ref: '#/definitions/summary.Totals'
        type: object
      date:
        type: string
      goal_source:
        type: string
      phase_name:
        type: string
      totals:
        $ref: '#/definitions/summary.Totals'
    type: object
  summary.Rolling:
    properties:
      adherence:
        $ref: '#/definitions/summary.Adherence'
      anchor_date:
        type: string
      averages:
        $ref: '#/definitions/summary.Totals'
      days:
        items:
          $ref: '#/definitions/summary.RollingDay'
        type: array
      days_with_data:
        type: integer
      goal_source:
        type: string
      phase_name:
        type: string
      total_days:
        type: integer
      tz:
        type: string
      window_days:
        type: integer
    type: object
  summary.RollingDay:
    properties:
      date:
        type: string
      has_data:
        type: boolean
      totals:
        $ref: '#/definitions/summary.Totals'
    type: object
  summary.Totals:
    properties:
      calcium_mg:
        type: number
      carbs_g:
        type: number
      fat_g:
        type: number
      fiber_g:
        type: number
      iron_mg:
        type: number
      kcal:
        type: number
      magnesium_mg:
        type: number
      potassium_mg:
        type: number
      protein_g:
        type: number
      salt_g:
        type: number
      sugar_g:
        type: number
      vitamin_b12_mcg:
        type: number
      vitamin_c_mg:
        type: number
      vitamin_d_mcg:
        type: number
      zinc_mg:
        type: number
    type: object
...
info:
  contact: {}
  description: Kazper — personal endurance-fueling and training-coach REST API. All
    endpoints under /products, /meals, and /summary require a bearer token.
  title: Kazper
  version: 0.1.0
paths:
...
  /summary/daily:
    get:
      description: Returns aggregated totals for a single calendar date in the supplied
        timezone. When `meal_type` is set, only entries of that type are summed and
        adherence is omitted.
      parameters:
      - description: Calendar date in YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: IANA timezone (defaults to DEFAULT_USER_TZ)
        in: query
        name: tz
        type: string
      - description: 'Optional filter: breakfast|lunch|dinner|snack'
        in: query
        name: meal_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/summary.Daily'
        "400":
          description: date_invalid | tz_invalid | meal_type_invalid
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Daily nutrition totals
      tags:
      - summary
...
  /summary/protein-distribution:
    get:
      description: 'Returns one row per `meal_entries` row on `date`, annotated with
        `mps_effective: bool` (against the 0.3 g/kg body-weight muscle-protein-synthesis
        threshold) plus `logged_at_hour` and `gap_minutes_since_previous` for circadian
        + meal-spacing context. The headline metric is `mps_effective_meal_count /
        meal_count`. Body weight resolution order: explicit `body_weight_kg` query
        param > rolling 7-day mean of stored entries ending at `date` (inclusive)
        > most-recent stored entry strictly before `date`. With no stored data and
        no override, returns `400 weight_data_missing`.'
      parameters:
      - description: Calendar date in YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: IANA timezone (defaults to DEFAULT_USER_TZ)
        in: query
        name: tz
        type: string
      - description: Explicit body weight override; > 0
        in: query
        name: body_weight_kg
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/summary.ProteinDistribution'
        "400":
          description: date_required | date_invalid | tz_invalid | body_weight_kg_invalid
            | weight_data_missing
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Per-meal protein distribution with MPS-threshold annotations
      tags:
      - summary
  /summary/range:
    get:
      description: Returns one row per day in `[from, to]`. Range is capped at 92
        days inclusive.
      parameters:
      - description: Inclusive start date YYYY-MM-DD
        in: query
        name: from
        required: true
        type: string
      - description: Inclusive end date YYYY-MM-DD
        in: query
        name: to
        required: true
        type: string
      - description: IANA timezone (defaults to DEFAULT_USER_TZ)
        in: query
        name: tz
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/summary.Range'
        "400":
          description: range_required | range_invalid | range_too_large | date_invalid
            | tz_invalid
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Daily totals across a date range
      tags:
      - summary
  /summary/rolling:
    get:
      description: 'Returns the trailing-window average of nutrition totals as of
        `anchor_date`. The window is `[anchor_date − (window_days − 1) days, anchor_date]`,
        both inclusive, in the requested `tz`. **Averages are computed across days
        with logged meals (`days_with_data`), NOT across `total_days`** — a 7-day
        window with 5 logged days returns the 5-day mean, with both divisors exposed
        so a sparse window is loud. Per-day rows carry `has_data: bool` so "no meals
        logged" stays distinct from "logged zero." Adherence is computed against the
        goal resolved at `anchor_date` (honoring per-date overrides). `window_days`
        is bounded `[2, 30]`.'
      parameters:
      - description: Calendar date in YYYY-MM-DD (the trailing window ends here)
        in: query
        name: anchor_date
        required: true
        type: string
      - description: Window size in calendar days; range [2, 30]
        in: query
        name: window_days
        required: true
        type: integer
      - description: IANA timezone (defaults to DEFAULT_USER_TZ)
        in: query
        name: tz
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/summary.Rolling'
        "400":
          description: anchor_date_required | anchor_date_invalid | window_days_required
            | window_days_invalid | tz_invalid
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Rolling-window nutrition average
      tags:
      - summary
...
securityDefinitions:
  BearerAuth:
    description: Static bearer token. Use the format `Bearer <token>` with either
      MOBILE_API_TOKEN or AGENT_API_TOKEN.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"